Removed plot function to allow users use desired projection and separate cutplotter and sliceplotter presenters implemented for cli #404
