+++ mslice/cli/__init__.py
@@ -2,13 +2,24 @@ from __future__ import (absolute_import, division, print_function)
 
 from ._mslice_commands import *  # noqa: F401
 
+
 import mantid.kernel
-import mantid.plots.plotfunctions
-import mantid.plots.plotfunctions3D
+from mslice.models.workspacemanager.workspace_provider import workspace_exists
+from mslice.plotting.cli_helperfunctions import is_slice, is_cut
+from mslice.workspace.histogram_workspace import HistogramWorkspace
+from mslice.workspace.base import WorkspaceBase as Workspace
+from mslice.workspace.workspace import Workspace as MatrixWorkspace
+from mslice.presenters.cut_plotter_presenter import CutPlotterPresenter
+from mslice.presenters.slice_plotter_presenter import SlicePlotterPresenter
+
+# Imports for mslice projections
 from matplotlib.axes import Axes
 from matplotlib.projections import register_projection
-from mslice.plotting.cli_helperfunctions import validate_args, is_cut
-from mslice.workspace.histogram_workspace import HistogramWorkspace
+
+
+# Separate cutplotter for cli
+CLI_CUT_PLOTTER_PRESENTER = CutPlotterPresenter()
+CLI_SLICE_PLOTTER_PRESENTER = SlicePlotterPresenter()
 
 
 def _check_workspace_name(workspace):
@@ -43,10 +54,16 @@ def PlotSlice(InputWorkspace, IntensityStart="", IntensityEnd="", Colormap=DEFAU
     :param Colormap: Colormap name as a string. Default is 'viridis'.
     :return:
     """
+    global CLI_SLICE_PLOTTER_PRESENTER
     _check_workspace_name(InputWorkspace)
     workspace = get_workspace_handle(InputWorkspace)
     _check_workspace_type(workspace, HistogramWorkspace)
-    slice_presenter = app.MAIN_WINDOW.slice_plotter_presenter
+    slice_presenter = CLI_SLICE_PLOTTER_PRESENTER
+
+    # slice cache needed from main slice plotter presenter
+    slice_presenter._slice_cache = app.MAIN_WINDOW.slice_plotter_presenter._slice_cache
+
+
     slice_presenter.change_intensity(workspace.name, IntensityStart, IntensityEnd)
     slice_presenter.change_colourmap(workspace.name, Colormap)
     slice_presenter.plot_from_cache(workspace)
@@ -63,6 +80,7 @@ def PlotCut(InputWorkspace, IntensityStart=0, IntensityEnd=0, PlotOver=False):
     :param PlotOver: if true the cut will be plotted on an existing figure.
     :return:
     """
+    global CLI_CUT_PLOTTER_PRESENTER
     _check_workspace_name(InputWorkspace)
     workspace = get_workspace_handle(InputWorkspace)
     if not isinstance(workspace, HistogramWorkspace):
@@ -71,7 +89,7 @@ def PlotCut(InputWorkspace, IntensityStart=0, IntensityEnd=0, PlotOver=False):
         intensity_range = None
     else:
         intensity_range = (IntensityStart, IntensityEnd)
-    app.MAIN_WINDOW.cut_plotter_presenter.plot_cut_from_workspace(workspace, intensity_range=intensity_range,
+    CLI_CUT_PLOTTER_PRESENTER.plot_cut_from_workspace(workspace, intensity_range=intensity_range,
                                                                   plot_over=PlotOver)
 
 
@@ -80,14 +98,25 @@ class MSliceAxes(Axes):
 
     def plot(self, *args, **kwargs):
 
-        if validate_args(*args):
-            mantid.kernel.logger.debug('using mantid.plots.plotfunctions')
-            if is_cut(*args):
+        if is_cut(*args):
+            try:
+                mantid.kernel.logger.debug('using mantid.plots.plotfunctions')
                 PlotCut(*args, **kwargs)
-            else:
-                PlotSlice(*args, **kwargs)
+            except Exception as e:
+                print('Mslice Projection Error: ' + repr(e))
         else:
             return Axes.plot(self, *args, **kwargs)
+
+    def pcolormesh(self, *args, **kwargs):
+
+        if is_slice(*args):
+            try:
+                mantid.kernel.logger.debug('using mantid.plots.plotfunctions')
+                PlotSlice(*args, **kwargs)
+            except Exception as e:
+                print('MSlice Projection Error: ' + repr(e))
+        else:
+            return Axes.pcolormesh(self, *args, **kwargs)
 
 
 register_projection(MSliceAxes)
